refactor: develop 이상한 부분 수정
