+++ src/pages/login/index.tsx
@@ -10,7 +10,7 @@ import Input from '@/components/common/input'
 import CommonButton from '@/components/common/commonbutton/CommonButton'
 import { authService } from '../../api/services/authServices'
 import { useAuthStore } from '@/stores/auth'
-import Modal from '@/components/modal/Modal'
+import Modal from '@/components/domain/modals/Modal'
 import { useFormSignup } from '@/hooks/useFormSignup'
 
 export default function Login() {
